// src/types/color.ts
export enum ProductColor {
  BLUE = "Синий",
  RED = "Красный",
  GREEN = "Зеленый",
  LIGHTBLUE = "Голубой",
  PURPLE = "Фиолетовый",
  YELLOW = "Желтый",
  BLACK = "Черный",
  WHITE = "Белый",
  PINK = "Розовый",
  GRAY = "Серый",
}